Migrated all animations to _animation folder.
