+++ src/utils.js
@@ -1,110 +1,6 @@
 import { format } from "date-fns";
 import { colors } from "./data/data";
-import gsap from "gsap";
-
-/* gsap timeline */
-let stagger = 0.05;
-let duration = 0.1;
-let duration2 = 0.2;
-let loginTL;
-let registerTL;
-let login;
-let register;
-let lefties;
-let righties;
-let ease = "power1.out";
-
-window.onload = () => {
-  lefties = document.querySelectorAll(".login-modal__left");
-  righties = document.querySelectorAll(".register-modal__right");
-  login = document.getElementsByClassName("login-modal--box-1");
-  register = document.getElementsByClassName("register-modal--box-2");
-
-  loginTL = gsap
-    .timeline()
-    .to(lefties, {
-      width: window.innerWidth > 375 ? "375px" : "100vw",
-      stagger: stagger,
-      duration: duration,
-      ease: ease,
-    })
-    .to(login, { opacity: 1, duration: duration2 });
-
-  registerTL = gsap
-    .timeline()
-    .to(righties, {
-      width: window.innerWidth > 375 ? "375px" : "100vw",
-      stagger: stagger,
-      duration: duration,
-      ease: ease,
-    })
-    .to(register, { opacity: 1, duration: duration2 });
-
-  loginTL.pause();
-  registerTL.pause();
-};
-
-window.addEventListener("resize", () => {
-  if (window.innerWidth > 375) {
-    if (loginTL && registerTL) {
-      loginTL.kill();
-      registerTL.kill();
-    }
-
-    loginTL = gsap
-      .timeline()
-      .to(lefties, {
-        width: "375px",
-        stagger: stagger,
-        duration: duration,
-        ease: ease,
-      })
-      .to(login, { opacity: 1, duration: duration2 });
-
-    registerTL = gsap
-      .timeline()
-      .to(righties, {
-        width: "375px",
-        stagger: stagger,
-        duration: duration,
-        ease: ease,
-      })
-      .to(register, { opacity: 1, duration: duration2 });
-
-    loginTL.pause();
-    registerTL.pause();
-  } else {
-    if (loginTL && registerTL) {
-      loginTL.kill();
-      registerTL.kill();
-    }
-
-    loginTL = gsap
-      .timeline()
-      .to(lefties, {
-        width: "100vw",
-        stagger: stagger,
-        duration: duration,
-        ease: ease,
-      })
-      .to(login, { opacity: 1, duration: duration2 });
-
-    registerTL = gsap
-      .timeline()
-      .to(righties, {
-        width: "100vw",
-        stagger: stagger,
-        duration: duration,
-        ease: ease,
-      })
-      .to(register, { opacity: 1, duration: duration2 });
-
-    loginTL.pause();
-    registerTL.pause();
-  }
-});
-
-/* end gsap timeline */
+import { drawerAnimations } from "./components/_animations/drawer.animations";
 
 const formatDate = (date) => {
   let createdAt = date.slice(0, 10).split("-");
@@ -128,7 +24,7 @@ const userIsLogged = (action) => {
   if (localStorage.getItem("user")) {
     action();
   } else {
-    loginTL.play();
+    drawerAnimations.loginPlay();
   }
 };
 
@@ -140,42 +36,4 @@ const loggedIn = () => {
   }
 };
 
-const loginPlay = () => {
-  loginTL.play();
-};
-
-const loginReverse = () => {
-  loginTL.reverse();
-};
-
-const registerPlay = () => {
-  registerTL.play();
-};
-
-const registerReverse = () => {
-  registerTL.reverse();
-};
-
-const play = () => {
-  loginPlay();
-  registerPlay();
-};
-
-const reverse = () => {
-  loginTL.reverse();
-  registerTL.reverse();
-};
-
-export {
-  formatDate,
-  getColor,
-  redirect,
-  userIsLogged,
-  loggedIn,
-  play,
-  reverse,
-  loginPlay,
-  registerPlay,
-  loginReverse,
-  registerReverse,
-};
+export { formatDate, getColor, redirect, userIsLogged, loggedIn };
